Решение 2.3 Изменить MealRepository и InMemoryMealRepositoryImpl таким образом, чтобы вся еда всех пользователей находилась в одном общем хранилище, но при этом каждый конкретный авторизованный пользователь мог видеть и редактировать только свою еду:  1) реализовать хранение еды для каждого пользователя можно с добавлением поля userId в Meal или без него (как вам больше нравится). Помните, что репозиторий один и приложение может работать одновременно со многими пользователями  2) если по запрошенному id еда отсутствует или чужая, возвращать null/false (в этом смысле тут всё аналогично UserRepository)  3) список еды возвращать отсортированным в обратном порядке по датам  4) атомарность операций не требуется (коллизии при одновременном изменении одного и того же пользователя можно не учитывать)

+++ src/main/java/org/zcorp/java2/repository/mock/InMemoryMealRepositoryImpl.java
@@ -1,45 +1,64 @@
 package org.zcorp.java2.repository.mock;
 
+import org.springframework.stereotype.Repository;
 import org.zcorp.java2.model.Meal;
 import org.zcorp.java2.repository.MealRepository;
 import org.zcorp.java2.util.MealsUtil;
 
-import java.util.Collection;
+import java.util.Comparator;
+import java.util.List;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.atomic.AtomicInteger;
 
+import static java.util.stream.Collectors.toList;
+
+@Repository
 public class InMemoryMealRepositoryImpl implements MealRepository {
     private Map<Integer, Meal> repository = new ConcurrentHashMap<>();
     private AtomicInteger counter = new AtomicInteger(0);
 
     {
-        MealsUtil.MEALS.forEach(this::save);
+        MealsUtil.MEALS.forEach(meal -> save(meal, meal.getUserId()));
     }
 
     @Override
-    public Meal save(Meal meal) {
+    public Meal save(Meal meal, int userId) {
+        if (meal.getUserId() != userId) {
+            return null;
+        }
         if (meal.isNew()) {
             meal.setId(counter.incrementAndGet());
             repository.put(meal.getId(), meal);
             return meal;
         }
-        // TODO Обработать ситуацию, когда мы попали сюда, т.е. на обновление, а метод computeIfPresent возвращает null (не находит id в repository)
         return repository.computeIfPresent(meal.getId(), (id, oldMeal) -> meal);
     }
 
     @Override
-    public void delete(int id) {
+    public boolean delete(int id, int userId) {
+        Meal meal = repository.get(id);
+        if (meal == null || meal.getUserId() != userId) {
+            return false;
+        }
         repository.remove(id);
+        return true;
     }
 
     @Override
-    public Meal get(int id) {
-        return repository.get(id);
+    public Meal get(int id, int userId) {
+        Meal meal = repository.get(id);
+        if (meal == null || meal.getUserId() != userId) {
+            return null;
+        }
+        return meal;
     }
 
     @Override
-    public Collection<Meal> getAll() {
-        return repository.values();
+    public List<Meal> getAll(int userId) {
+        return repository.values().stream()
+                .filter(meal -> meal.getUserId() == userId)
+                .sorted(Comparator.comparing(Meal::getDateTime).reversed())
+                .collect(toList());
     }
 }
